+++ tests/test_password_auth_succeeds_with_correct_credentials.c
@@ -14,7 +14,8 @@ int test(LIBSSH2_SESSION *session)
     int rc;
 
     const char *userauth_list =
-        libssh2_userauth_list(session, USERNAME, strlen(USERNAME));
+        libssh2_userauth_list(session, USERNAME,
+                              (unsigned int)strlen(USERNAME));
     if(userauth_list == NULL) {
         print_last_session_error("libssh2_userauth_list");
         return 1;
@@ -26,8 +27,10 @@ int test(LIBSSH2_SESSION *session)
         return 1;
     }
 
-    rc = libssh2_userauth_password_ex(session, USERNAME, strlen(USERNAME),
-                                      PASSWORD, strlen(PASSWORD), NULL);
+    rc = libssh2_userauth_password_ex(session, USERNAME,
+                                      (unsigned int)strlen(USERNAME),
+                                      PASSWORD,
+                                      (unsigned int)strlen(PASSWORD), NULL);
     if(rc != 0) {
         print_last_session_error("libssh2_userauth_password_ex");
         return 1;
